+++ app/controllers/documents_controller.rb
@@ -4,7 +4,7 @@
 class DocumentsController < ApplicationController
   include DocumentListing
 
-  before_action :set_document, only: %i[show download]
+  before_action :set_document, only: %i[show preview download]
 
   def index
     load_documents_list
@@ -17,9 +17,27 @@ class DocumentsController < ApplicationController
       redirect_to documents_path, alert: t("views.documents.file_missing")
       return
     end
+  end
 
-    @inline_url = rails_blob_path(@document.file, disposition: "inline")
-    @download_url = rails_blob_path(@document.file, disposition: "attachment")
+  def preview
+    authorize @document, :show?
+
+    ocr_log = @document.latest_searchable_ocr_log
+    if ocr_log&.searchable_pdf&.attached?
+      blob = ocr_log.searchable_pdf.blob
+      send_data blob.download,
+                filename: blob.filename.to_s,
+                type: "application/pdf",
+                disposition: "inline"
+    elsif @document.file.attached?
+      blob = @document.file.blob
+      send_data blob.download,
+                filename: blob.filename.to_s,
+                type: "application/pdf",
+                disposition: "inline"
+    else
+      redirect_to documents_path, alert: t("views.documents.file_missing")
+    end
   end
 
   def download