+++ config/routes.rb
@@ -31,6 +31,7 @@ Rails.application.routes.draw do
   resources :documents, only: [:index, :show] do
     member do
       get :download
+      get :preview
     end
   end